[SPARK-39936][SQL] Store schema in properties for Spark Views
We normally store the tableSchema in table properties rather than as an actual schema. We normally do this as it helps bypass some hive-metastore issues including the parsing with the SparkSQLDataType parser.

However, the issue is that we were not emptying out the tableSchema for spark views (which are incompatible with Hive to start with). Thus, the hive DataType parser would try to parse the schema and it would break. The fix is to empty out the tableSchema for spark views when saving them to Hive Metastore.

This fixes the following bug:
```
-- this works since we use backticks in `date-of-creation`
create table table_with_hyphen (f array<struct<validColumnTypeName:string>>) using parquet;
-- this should work since we use backticks in `date-of-creation` but will break without this bugfix
create or replace view view_with_hyphen as select f from table_with_hyphen;
```

Bugfix as described above.

Tested in spark and added a unit test.

Closes #37364 from Jeffreychen99/SPARK-39936.

Authored-by: Jeffrey Chen <jeffrey.chen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>